[WIP] Refactoring - pdfanno.js
