+++ src/page/util/display.js
@@ -1,0 +1,136 @@
+
+
+
+
+
+export function displayAnnotation(isPrimary) {
+
+    let annotations = [];
+    let colors = [];
+    let primaryIndex = -1;
+
+    // Primary annotation.
+    if (isPrimary) {
+        $('#dropdownAnnoPrimary a').each((index, element) => {
+            let $elm = $(element);
+            if ($elm.find('.fa-check').hasClass('no-visible') === false) {
+                let annoPath = $elm.find('.js-annoname').text();
+                if (!fileMap[annoPath]) {
+                    console.log('ERROR');
+                    return;
+                }
+                primaryIndex = 0;
+                annotations.push(fileMap[annoPath]);
+                let color = null; // Use the default color used for edit.
+                colors.push(color);
+
+                let filename = annoPath.split('/')[annoPath.split('/').length - 1];
+                localStorage.setItem('_pdfanno_primary_annoname', filename);
+                console.log('filename:', filename);
+            }
+        });
+    }
+
+    // Reference annotations.
+    if (!isPrimary) {
+        $('#dropdownAnnoReference a').each((index, element) => {
+            let $elm = $(element);
+            if ($elm.find('.fa-check').hasClass('no-visible') === false) {
+                let annoPath = $elm.find('.js-annoname').text();
+                if (!fileMap[annoPath]) {
+                    console.log('ERROR');
+                    return;
+                }
+                annotations.push(fileMap[annoPath]);
+                let color = $elm.find('.js-anno-palette').spectrum('get').toHexString();
+                console.log(color);
+                colors.push(color);
+            }
+        });
+    }
+
+    console.log('colors:', colors);
+
+    // Create import data.
+    let paperData = {
+        primary : primaryIndex,
+        colors,
+        annotations
+    };
+
+    // Pass the data to pdf-annotatejs.
+    window.iframeWindow.PDFAnnoCore.getStoreAdapter().importAnnotations(paperData, isPrimary).then(result => {
+
+        // Reload the viewer.
+        reloadPDFViewer();
+
+        // Reset tools to viewMode.
+        $('.js-tool-btn[data-type="view"]').click();
+    });
+
+}
+
+
+/*
+var event = document.createEvent('CustomEvent');
+event.initCustomEvent('annotationrendered', true, true, null);
+window.dispatchEvent(event);
+*/
+
+
+
+/**
+ * Reload PDF Viewer.
+ */
+export function reloadPDFViewer() {
+
+    // Reload pdf.js.
+    $('#viewer iframe').remove();
+    $('#viewer').html('<iframe src="./pages/viewer.html" class="anno-viewer" frameborder="0"></iframe>');
+
+    // Restart.
+    var event = document.createEvent('CustomEvent');
+    event.initCustomEvent('restartApp', true, true, null);
+    window.dispatchEvent(event);
+}
+
+
+/**
+ * Clear the all annotations from the view and storage.
+ */
+export function clearAllAnnotations() {
+    localStorage.removeItem('_pdfanno_containers');
+    localStorage.removeItem('_pdfanno_primary_annoname');
+}
+
+/**
+ * Setup the color pickers.
+ */
+export function setupColorPicker() {
+
+    const colors = [
+        'rgb(255, 128, 0)', 'hsv 100 70 50', 'yellow', 'blanchedalmond',
+        'red', 'green', 'blue', 'violet'
+    ];
+
+    // Setup colorPickers.
+    $('.js-anno-palette').spectrum({
+        showPaletteOnly        : true,
+        showPalette            : true,
+        hideAfterPaletteSelect : true,
+        palette                : [
+            colors.slice(0, Math.floor(colors.length/2)),
+            colors.slice(Math.floor(colors.length/2), colors.length)
+        ]
+    });
+    // Set initial color.
+    $('.js-anno-palette').each((i, elm) => {
+        $(elm).spectrum('set', colors[ i % colors.length ]);
+    });
+
+    // Setup behavior.
+    $('.js-anno-palette').off('change').on('change', displayAnnotation.bind(null, false));
+}
+
+
+
